# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "New Voting System",
        "version": "1.0.0",
        "description": "API of the New Voting System. Conventions: docs/api/README.md."
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Same origin"
        }
    ],
    "paths": {
...
        "/v1/ballots/{ballot}": {
            "patch": {
                "operationId": "ballot.update",
                "description": "Every field is optional; only those sent change. Not the position (see the order\nendpoint). A draft election only. Owner or manager. Limited to 120 requests per hour per user.",
                "summary": "Change a ballot",
                "tags": [
                    "Ballot"
                ],
                "parameters": [
                    {
                        "name": "ballot",
                        "in": "path",
                        "required": true,
                        "description": "The ballot UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBallotRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "position": {
                                                    "type": "integer"
                                                },
                                                "seats": {
                                                    "type": "integer"
                                                },
                                                "allow_blank": {
                                                    "type": "boolean"
                                                },
                                                "candidates_count": {
                                                    "type": "integer"
                                                },
                                                "candidates": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string"
                                                            },
                                                            "ballot": {
                                                                "type": "string"
                                                            },
                                                            "party": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "first_name": {
                                                                "type": "string"
                                                            },
                                                            "last_name": {
                                                                "type": "string"
                                                            },
                                                            "sex": {
                                                                "type": "string"
                                                            },
                                                            "slogan": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "biography": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "photo": {
                                                                "type": [
                                                                    "object",
                                                                    "null"
                                                                ],
                                                                "properties": {
                                                                    "sm": {
                                                                        "type": "string"
                                                                    },
                                                                    "md": {
                                                                        "type": "string"
                                                                    }
                                                                },
                                                                "required": [
                                                                    "sm",
                                                                    "md"
                                                                ]
                                                            },
                                                            "position": {
                                                                "type": "integer"
                                                            },
                                                            "created_at": {
                                                                "type": "string"
                                                            },
                                                            "updated_at": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "required": [
                                                            "id",
                                                            "ballot",
                                                            "party",
                                                            "first_name",
                                                            "last_name",
                                                            "sex",
                                                            "slogan",
                                                            "biography",
                                                            "photo",
                                                            "position",
                                                            "created_at",
                                                            "updated_at"
                                                        ]
                                                    }
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "title",
                                                "description",
                                                "position",
                                                "seats",
                                                "allow_blank",
                                                "candidates_count",
                                                "candidates",
                                                "created_at",
                                                "updated_at"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "ballot.destroy",
                "description": "A draft election only. The positions after it close the gap. Owner or manager. Limited\nto 120 requests per hour per user.",
                "summary": "Delete a ballot",
                "tags": [
                    "Ballot"
                ],
                "parameters": [
                    {
                        "name": "ballot",
                        "in": "path",
                        "required": true,
                        "description": "The ballot UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/candidates/{candidate}": {
            "patch": {
                "operationId": "candidate.update",
                "description": "Every field is optional; only those sent change. `ballot` moves it to the end of another\nballot of the same election. Not the position (see the order endpoint) nor the photo. A\ndraft election only. Owner or manager. Limited to 120 requests per hour per user.",
                "summary": "Change a candidate",
                "tags": [
                    "Candidate"
                ],
                "parameters": [
                    {
                        "name": "candidate",
                        "in": "path",
                        "required": true,
                        "description": "The candidate UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCandidateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "ballot": {
                                                    "type": "string"
                                                },
                                                "party": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "first_name": {
                                                    "type": "string"
                                                },
                                                "last_name": {
                                                    "type": "string"
                                                },
                                                "sex": {
                                                    "type": "string",
                                                    "enum": [
                                                        "male",
                                                        "female"
                                                    ]
                                                },
                                                "slogan": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "biography": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "photo": {
                                                    "type": [
                                                        "object",
                                                        "null"
                                                    ],
                                                    "properties": {
                                                        "sm": {
                                                            "type": "string"
                                                        },
                                                        "md": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "sm",
                                                        "md"
                                                    ]
                                                },
                                                "position": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "ballot",
                                                "party",
                                                "first_name",
                                                "last_name",
                                                "sex",
                                                "slogan",
                                                "biography",
                                                "photo",
                                                "position",
                                                "created_at",
                                                "updated_at"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "candidate.destroy",
                "description": "A draft election only. The positions after it close the gap. Owner or manager. Limited to\n120 requests per hour per user.",
                "summary": "Delete a candidate",
                "tags": [
                    "Candidate"
                ],
                "parameters": [
                    {
                        "name": "candidate",
                        "in": "path",
                        "required": true,
                        "description": "The candidate UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/invitations/{invitation}": {
            "delete": {
                "operationId": "invitation.destroy",
                "description": "Its link stops working; the row is deleted. An invitation that is unknown, already\naccepted, already cancelled or another institution's answers 404. Owner only.",
                "summary": "Cancel an invitation",
                "tags": [
                    "Invitation"
                ],
                "parameters": [
                    {
                        "name": "invitation",
                        "in": "path",
                        "required": true,
                        "description": "The invitation UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/v1/users/{user}": {
            "delete": {
                "operationId": "user.destroy",
                "description": "The user can no longer sign in and their next request is refused. Their address is freed;\ntheir name is kept for the audit log. An institution keeps at least one owner. Removing\noneself also ends one's own session. Owner only.",
                "summary": "Remove a user",
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "description": "The user UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateBallotRequest": {
                "type": "object",
                "description": "docs/api/ballots/PATCH-ballots-{ballot}.md: the rules of the creation, every field optional.",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 200
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 1000
                    },
                    "seats": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 20
                    },
                    "allow_blank": {
                        "type": "boolean"
                    }
                },
                "title": "UpdateBallotRequest"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}